+++ skbuild/command/bdist_wheel.py
@@ -1,6 +1,7 @@
 """This module defines custom implementation of ``bdist_wheel`` setuptools
 command."""
 
+import os
 import sys
 
 try:
@@ -60,12 +61,12 @@ class bdist_wheel(set_build_base_mixin, new_style(_bdist_wheel)):
                 _wheel_archive.make_wheelfile_inner = old_make_wheelfile_inner
 
     def finalize_options(self, *args, **kwargs):
-        """Ensure MacOSX wheels include ``x86_64`` instead of ``intel``."""
+        """Ensure MacOSX wheels include the expected platform information."""
         # pylint:disable=attribute-defined-outside-init,access-member-before-definition
         if sys.platform == 'darwin' and self.plat_name is None and self.distribution.has_ext_modules():
-            # The following code is duplicated in setuptools_wrap
+            # The default value is duplicated in setuptools_wrap
             # pylint:disable=access-member-before-definition
-            self.plat_name = "macosx-10.6-x86_64"
+            self.plat_name = os.environ.get('_PYTHON_HOST_PLATFORM', 'macosx-10.6-x86_64')
         super(bdist_wheel, self).finalize_options(*args, **kwargs)
 
     def write_wheelfile(self, wheelfile_base, _=None):