+++ skbuild/command/egg_info.py
@@ -24,7 +24,7 @@ class egg_info(set_build_base_mixin, new_style(_egg_info)):
                 # running the test command.
                 package_name = list(self.distribution.package_dir.keys())[0]
                 egg_base = to_unix_path(list(self.distribution.package_dir.values())[0])
-                cmake_install_dir = to_unix_path(CMAKE_INSTALL_DIR)
+                cmake_install_dir = to_unix_path(CMAKE_INSTALL_DIR())
                 if egg_base.startswith(cmake_install_dir):
                     egg_base = egg_base[len(cmake_install_dir) + 1:]
                 if package_name and egg_base.endswith(package_name):